+++ src/index.ts
@@ -1,7 +1,9 @@
 import { ClientConfig, Credentials } from './core';
+import { Dictionaries } from './dictionaries';
 import { Distributions } from './distributions';
 import { Glossaries } from './glossaries';
 import { Issues } from './issues';
+import { Labels } from './labels';
 import { Languages } from './languages';
 import { MachineTranslation } from './machineTranslation';
 import { ProjectsGroups } from './projectsGroups';
@@ -22,9 +24,11 @@ import { Webhooks } from './webhooks';
 import { Workflows } from './workflows';
 
 export * from './core';
+export * from './dictionaries';
 export * from './distributions';
 export * from './glossaries';
 export * from './issues';
+export * from './labels';
 export * from './languages';
 export * from './machineTranslation';
 export * from './projectsGroups';
@@ -66,6 +70,8 @@ export default class Client {
     readonly issuesApi: Issues;
     readonly teamsApi: Teams;
     readonly distributionsApi: Distributions;
+    readonly dictionariesApi: Dictionaries;
+    readonly labelsApi: Labels;
 
     constructor(credentials: Credentials, config?: ClientConfig) {
         this.sourceFilesApi = new SourceFiles(credentials, config);
@@ -89,5 +95,7 @@ export default class Client {
         this.issuesApi = new Issues(credentials, config);
         this.teamsApi = new Teams(credentials, config);
         this.distributionsApi = new Distributions(credentials, config);
+        this.dictionariesApi = new Dictionaries(credentials, config);
+        this.labelsApi = new Labels(credentials, config);
     }
 }