+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to the Crowdin API Client will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.9.0]
+
+### Added
+
+- Dictionaries API
+- Labels API
+
+### Updated
+
+- Projects API (minor changes)
+- Files API (minor changes)
+- Strings API (minor changes)
+- String Translations API (minor changes)
+- Tasks API (minor changes)
+- Translations API (minor changes)
+- Users API (new methods)
+
 ## [1.8.0]
 
 ### Added